pyproject.toml
[project]
name = "calhacks"
version = "0.1.0"
description = "Python utilities for XRPL settlement graph logic."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "xrpl-py",
    "fastapi",
    "uvicorn",
    "pydantic",
    "pytest",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
]

[tool.setuptools]
packages = ["backend"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
